+++ binaries/pico_limbo/src/configuration/game_mode_config.rs
@@ -1,0 +1,11 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Deserialize, Serialize, Default)]
+#[serde(rename_all = "lowercase")]
+pub enum GameModeConfig {
+    Survival,
+    Creative,
+    Adventure,
+    #[default]
+    Spectator,
+}