+++ binaries/pico_limbo/src/configuration/mod.rs
@@ -1,5 +1,5 @@
 pub mod config;
 mod forwarding;
-mod game_mode;
+pub mod game_mode_config;
 mod server_list;
 mod spawn_dimension;