+++ binaries/pico_limbo/src/server/mod.rs
@@ -2,6 +2,7 @@ pub mod client;
 mod client_inner;
 mod controllable_interval;
 pub mod event_handler;
+pub mod game_mode;
 pub mod game_profile;
 mod named_packet;
 pub mod network;